[REFACTOR] 닉네임 검증과 프로필 이미지 타입 변수의 위치를 ui/common으로 변경
Fixes: TICO-434
Related to: TICO-432
