+++ Mobile/PomoroDo/app/src/main/java/com/tico/pomorodo/ui/common/view/NameErrorType.kt
@@ -1,4 +1,4 @@
-package com.tico.pomorodo.data.model
+package com.tico.pomorodo.ui.common.view
 
 enum class NameErrorType {
     NONE, INVALID_ERROR, RANGE_ERROR
